+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.4
+
+- water pathway
+
+## 0.0.3
+
+- edit button for property details
+
+## 0.0.2
+
+- instructions for upgrading npm to 18
+
 ## 0.0.1
 
 - first PR by fred